roles

+++ backend/app1/Views/FanViews.py
@@ -1,16 +1,19 @@
 from django.db.models import Avg
+from rest_framework.permissions import IsAuthenticatedOrReadOnly
 from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework import status, generics
 
 from .Pagination import CustomPagination
 from ..models import Fan, SwimmerFan
+from ..permissions import HasEditPermissionOrReadOnly
 from ..serailizer import FanSerializer, FanSerializerId, SwimmerFanSerializer, FanSerializerAvg
 
 
 class FanListCreateView(generics.ListCreateAPIView):
     serializer_class = FanSerializer
     pagination_class = CustomPagination
+    permission_classes = [IsAuthenticatedOrReadOnly]
 
     def get_queryset(self):
         queryset = Fan.objects.all()
@@ -30,6 +33,7 @@ class FanListCreateView(generics.ListCreateAPIView):
 class FanInfo(APIView):
     serializer_class = FanSerializerId
     pagination_class = CustomPagination
+    permission_classes = [IsAuthenticatedOrReadOnly, HasEditPermissionOrReadOnly]
 
     def get(self, request, id):
         try:
@@ -60,6 +64,8 @@ class FanInfo(APIView):
             msg = {"msg": "not found error"}
             return Response(msg, status=status.HTTP_404_NOT_FOUND)
 
+        self.check_object_permissions(request, obj)
+
         serializer = FanSerializerId(obj, data=request.data)
 
         if serializer.is_valid():
@@ -75,6 +81,8 @@ class FanInfo(APIView):
             msg = {"msg": "not found error"}
             return Response(msg, status=status.HTTP_404_NOT_FOUND)
 
+        self.check_object_permissions(request, obj)
+
         serializer = FanSerializer(obj, data=request.data, partial=True)
 
         if serializer.is_valid():
@@ -89,6 +97,8 @@ class FanInfo(APIView):
         except Fan.DoesNotExist:
             msg = {"msg": "not found"}
             return Response(msg, status=status.HTTP_404_NOT_FOUND)
+
+        self.check_object_permissions(request, obj)
 
         obj.delete()
         return Response({"msg": "deleted"}, status=status.HTTP_204_NO_CONTENT)
